ai/requirements.txt
Flask 
numpy 
faiss-cpu 
sentence-transformers 
google-generativeai
requests  
python-dotenv